pyproject.toml
[project]
name = "tamago"
version = "0.1.0"
description = "自分の分身AIを育てるCLIツール"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Kazuaki Yokura (U73)" }]
dependencies = [
    "typer>=0.9.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
ollama = ["ollama>=0.1.0"]
llamacpp = ["llama-cpp-python>=0.3.0"]
all = ["openai>=1.0.0", "ollama>=0.1.0", "llama-cpp-python>=0.3.0"]

[project.scripts]
tamago = "tamago.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tamago"]
